requirements.txt
requests
pandas
urllib3
logging
python-dotenv